feat: split drive & user class + utils functions
