+++ src/utils/arraytbufferToString.ts
@@ -1,0 +1,1 @@
+export const ArraybufferToString = (ab: ArrayBuffer): string => new TextDecoder().decode(ab);
