+++ src/utils/mnemonicToPrivateKey.ts
@@ -1,0 +1,3 @@
+import * as ethers from 'ethers';
+
+export const mnemonicToPrivateKey = (mnemonic: string): string => ethers.Wallet.fromMnemonic(mnemonic).privateKey;
